/register, /myprofile 디버깅 완료 (1,3,4 완료)

+++ main/calc/EXPLevelFormula.py
@@ -2,4 +2,4 @@ import numpy
 
 
 def minEXP(level: int):
-    return numpy.ceil(numpy.divide(numpy.power(1.2, level)-1.2, 0.012))
+    return int(numpy.ceil(numpy.divide(numpy.power(1.2, level)-1.2, 0.012)))
